Move view factory to separate file
