+++ ios/FlutterShowView/FlutterShowView.swift
@@ -8,32 +8,6 @@ import Flutter
 import UIKit
 import VGSShowSDK
 
-class FlutterShowViewFactory: NSObject, FlutterPlatformViewFactory {
-
-	// MARK: - Private vars
-
-	private var messenger: FlutterBinaryMessenger
-
-  // MARK: - Initialization
-
-	init(messenger: FlutterBinaryMessenger) {
-		self.messenger = messenger
-	}
-
-	// MARK: - Public
-
-	public func create(withFrame frame: CGRect,
-										 viewIdentifier viewId: Int64,
-										 arguments args: Any?) -> FlutterPlatformView {
-		return FlutterShowView(messenger: messenger,
-															frame: frame, viewId: viewId,
-															args: args)
-	}
-	public func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {
-		return FlutterStandardMessageCodec.sharedInstance()
-	}
-}
-
 class FlutterShowView: NSObject, FlutterPlatformView {
 
 	let showView: ShowView
